Update 2816. Double a Number Represented as a Linked List.c

+++ C/2816. Double a Number Represented as a Linked List.c
@@ -14,7 +14,7 @@ void dfs(struct ListNode* node) {
     int temp = node->val * 2 + carry;
 
     node->val = temp % 10;
-    carry = (int)(temp / 10);
+    carry = temp / 10;
 }
 
 struct ListNode* doubleIt(struct ListNode* head) {
